+++ resources/views/components/header.blade.php
@@ -12,12 +12,12 @@
             <a href="#" data-toggle="dropdown" class="nav-link dropdown-toggle nav-link-lg nav-link-user">
                 <img alt="image" src="{{ asset('img/avatar/avatar-1.png') }}" class="rounded-circle mr-1">
                 <div class="d-sm-none d-lg-inline-block">
-                    Hai, {{ substr(auth()->user()->name, 0, 10) }}
+                    Hi, {{ substr(auth()->user()->name, 0, 10) }}
                 </div>
             </a>
             <div class="dropdown-menu dropdown-menu-right">
                 <div class="dropdown-title">
-                    Selamat Datang, {{ substr(auth()->user()->name, 0, 10) }}
+                    Welcome, {{ substr(auth()->user()->name, 0, 10) }}
                 </div>
                 <a class="dropdown-item has-icon edit-profile" href="{{ route('profile.edit') }}">
                     <i class="fa fa-user"></i> Edit Profile